Move webview out of notebooks
