+++ src/interactive-window/editor-integration/cellFactory.ts
@@ -6,12 +6,13 @@ import '../../platform/common/extensions';
 import type * as nbformat from '@jupyterlab/nbformat';
 import { NotebookCell, NotebookCellKind, NotebookDocument, Range, TextDocument, Uri } from 'vscode';
 
-import { appendLineFeed, generateMarkdownFromCodeLines, parseForComments } from '../../webviews/webview-side/common';
+import { appendLineFeed, generateMarkdownFromCodeLines } from '../../webviews/webview-side/common';
 import { CellMatcher } from './cellMatcher';
 import { ICell, ICellRange, IJupyterSettings, Resource } from '../../platform/common/types';
 import { noop } from '../../platform/common/utils/misc';
 import { getInteractiveCellMetadata } from '../helpers';
 import { createJupyterCellFromVSCNotebookCell } from '../../kernels/execution/helpers';
+import { parseForComments } from '../../platform/common/utils';
 
 export function createCodeCell(): nbformat.ICodeCell;
 // eslint-disable-next-line @typescript-eslint/unified-signatures
