+++ src/platform/common/utils.ts
@@ -273,3 +273,86 @@ export function formatStreamText(str: string): string {
     // Do the same thing jupyter is doing
     return fixCarriageReturn(fixBackspace(str));
 }
+
+const SingleQuoteMultiline = "'''";
+const DoubleQuoteMultiline = '"""';
+
+// eslint-disable-next-line complexity
+export function parseForComments(
+    lines: string[],
+    foundCommentLine: (s: string, i: number) => void,
+    foundNonCommentLine: (s: string, i: number) => void
+) {
+    // Check for either multiline or single line comments
+    let insideMultilineComment: string | undefined;
+    let insideMultilineQuote: string | undefined;
+    let pos = 0;
+    for (const l of lines) {
+        const trim = l.trim();
+        // Multiline is triple quotes of either kind
+        const isMultilineComment = trim.startsWith(SingleQuoteMultiline)
+            ? SingleQuoteMultiline
+            : trim.startsWith(DoubleQuoteMultiline)
+            ? DoubleQuoteMultiline
+            : undefined;
+        const isMultilineQuote = trim.includes(SingleQuoteMultiline)
+            ? SingleQuoteMultiline
+            : trim.includes(DoubleQuoteMultiline)
+            ? DoubleQuoteMultiline
+            : undefined;
+
+        // Check for ending quotes of multiline string
+        if (insideMultilineQuote) {
+            if (insideMultilineQuote === isMultilineQuote) {
+                insideMultilineQuote = undefined;
+            }
+            foundNonCommentLine(l, pos);
+            // Not inside quote, see if inside a comment
+        } else if (insideMultilineComment) {
+            if (insideMultilineComment === isMultilineComment) {
+                insideMultilineComment = undefined;
+            }
+            if (insideMultilineComment) {
+                foundCommentLine(l, pos);
+            }
+            // Not inside either, see if starting a quote
+        } else if (isMultilineQuote && !isMultilineComment) {
+            // Make sure doesn't begin and end on the same line.
+            const beginQuote = trim.indexOf(isMultilineQuote);
+            const endQuote = trim.lastIndexOf(isMultilineQuote);
+            insideMultilineQuote = endQuote !== beginQuote ? undefined : isMultilineQuote;
+            foundNonCommentLine(l, pos);
+            // Not starting a quote, might be starting a comment
+        } else if (isMultilineComment) {
+            // See if this line ends the comment too or not
+            const endIndex = trim.indexOf(isMultilineComment, 3);
+            insideMultilineComment = endIndex >= 0 ? undefined : isMultilineComment;
+
+            // Might end with text too
+            if (trim.length > 3) {
+                foundCommentLine(trim.slice(3, endIndex >= 0 ? endIndex : undefined), pos);
+            }
+        } else {
+            // Normal line
+            if (trim.startsWith('#')) {
+                foundCommentLine(trim.slice(1), pos);
+            } else {
+                foundNonCommentLine(l, pos);
+            }
+        }
+        pos += 1;
+    }
+}
+
+// Strip out comment lines from code
+export function stripComments(str: string): string {
+    let result: string = '';
+    parseForComments(
+        str.splitLines({ trim: false, removeEmptyEntries: false }),
+        (_s) => {
+            // Do nothing
+        },
+        (s) => (result = result.concat(`${s}\n`))
+    );
+    return result;
+}
