+++ src/test/datascience/cellFactory.unit.test.ts
@@ -2,9 +2,10 @@
 // Licensed under the MIT License.
 'use strict';
 import { assert } from 'chai';
-import { removeLinesFromFrontAndBack, stripComments } from '../../webviews/webview-side/common';
+import { removeLinesFromFrontAndBack } from '../../webviews/webview-side/common';
 import * as vscode from 'vscode';
 import { generateCells } from '../../interactive-window/editor-integration/cellFactory';
+import { stripComments } from '../../platform/common/utils';
 
 /* eslint-disable  */
 suite('DataScience CellFactory', () => {
