+++ src/webviews/extension-side/plotView/rendererCommunication.node.ts
@@ -3,10 +3,10 @@
 
 import { inject, injectable } from 'inversify';
 import { Event, extensions, NotebookEditor, window } from 'vscode';
-import { IExtensionSingleActivationService } from '../../platform/activation/types';
-import { disposeAllDisposables } from '../../platform/common/helpers';
-import { IDisposable } from '../../platform/common/types';
-import { noop } from '../../platform/common/utils/misc';
+import { IExtensionSingleActivationService } from '../../../platform/activation/types';
+import { disposeAllDisposables } from '../../../platform/common/helpers';
+import { IDisposable } from '../../../platform/common/types';
+import { noop } from '../../../platform/common/utils/misc';
 import { PlotSaveHandler } from './plotSaveHandler.node';
 import { PlotViewHandler } from './plotViewHandler.node';
